+++ gui/resources/hand_icon.png
